--- packages/code-analyzer-pmd-engine/pmd-rules/src/test/resources/sfca/rulesets/appexchange_apex/xml/AvoidSecurityEnforcedOldApiVersion.xml
@@ -1,1 +1,0 @@
-<!-- todo: can I make a test for this? -->
