+++ packages/code-analyzer-pmd-engine/test/test-data/pmdGoldfiles/rules_allLanguages.goldfile.json
@@ -259,7 +259,7 @@
   },
   {
     "name": "AvoidChangeProtectionUnprotected",
-    "severityLevel": 2,
+    "severityLevel": 1,
     "tags": [
       "AppExchange",
       "Security",
